[PLAYER-5077] Removed extra view
